+++ home/import_content_pages.py
@@ -40,9 +40,17 @@ class ImportException(Exception):
     Base exception for all import related issues.
     """
 
-    def __init__(self, message: str, row_num: int | None = None):
+    def __init__(
+        self,
+        message: str,
+        row_num: int | None = None,
+        slug: str | None = None,
+        locale: Locale | None = None,
+    ):
         self.row_num = row_num
         self.message = message
+        self.slug = slug
+        self.locale = locale
         super().__init__()
 
 
@@ -117,6 +125,8 @@ class ContentImporter:
                     self.add_message_to_shadow_content_page_from_row(row, prev_locale)
             except ImportException as e:
                 e.row_num = i
+                e.slug = row.slug
+                e.locale = row.locale
                 raise e
 
     def save_pages(self) -> None:
@@ -281,8 +291,9 @@ class ContentImporter:
         if len(row.footer) > 60:
             raise ImportException(f"footer too long: {row.footer}", row.page_id)
 
-        if len(row.list_items) > 24:
-            raise ImportException(f"list_items too long: {row.list_items}", row.page_id)
+        for list_item in row.list_items:
+            if len(list_item) > 24:
+                raise ImportException(f"list_items too long: {list_item}", row.page_id)
 
         self.shadow_pages[(row.slug, locale)] = page
 
@@ -316,20 +327,15 @@ class ContentImporter:
         if row.translation_tag or locale != self.default_locale():
             page.translation_key = row.translation_tag
 
-    def find_shadow_content_page(
-        self, row: "ContentRow", locale: Locale
-    ) -> "ShadowContentPage":
-        try:
-            return self.shadow_pages[(row.slug, locale)]
-        except KeyError:
-            raise ImportException(
-                f"Cannot find content page with slug '{row.slug}' and locale '{locale}'"
-            )
-
     def add_variation_to_shadow_content_page_from_row(
         self, row: "ContentRow", locale: Locale
     ) -> None:
-        page = self.find_shadow_content_page(row, locale)
+        try:
+            page = self.shadow_pages[(row.slug, locale)]
+        except KeyError:
+            raise ImportException(
+                f"This is a variation for the content page with slug '{row.slug}' and locale '{locale}', but no such page exists"
+            )
         whatsapp_block = page.whatsapp_body[-1]
         whatsapp_block.variation_messages.append(
             ShadowVariationBlock(
@@ -340,7 +346,12 @@ class ContentImporter:
     def add_message_to_shadow_content_page_from_row(
         self, row: "ContentRow", locale: Locale
     ) -> None:
-        page = self.find_shadow_content_page(row, locale)
+        try:
+            page = self.shadow_pages[(row.slug, locale)]
+        except KeyError:
+            raise ImportException(
+                f"This is a message for page with slug '{row.slug}' and locale '{locale}', but no such page exists"
+            )
         if row.is_whatsapp_message:
             page.enable_whatsapp = True
             buttons = []