+++ home/tests/test_models.py
@@ -1,13 +1,17 @@
 from io import StringIO
 from unittest import mock
 
+import pytest
 from django.core.exceptions import ValidationError
 from django.core.management import call_command
 from django.test import TestCase, override_settings
 from requests import HTTPError
 from wagtail.blocks import StructBlockValidationError
 from wagtail.images import get_image_model
-from wagtail.models import Page
+from wagtail.models import (
+    Locale,  # type: ignore
+    Page,
+)
 from wagtail.test.utils import WagtailPageTests
 
 from home.models import (
@@ -23,7 +27,7 @@ from home.models import (
     WhatsappBlock,
 )
 
-from .page_builder import PageBuilder, WABlk, WABody
+from .page_builder import PageBtn, PageBuilder, WABlk, WABody
 from .utils import create_page, create_page_rating
 
 
@@ -79,10 +83,12 @@ class ContentPageTests(TestCase):
     @mock.patch("home.models.create_whatsapp_template")
     def test_template_create_on_save(self, mock_create_whatsapp_template):
         page = create_page(is_whatsapp_template=True)
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_with(
             f"wa_title_{page.get_latest_revision().id}",
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             [],
             None,
             [],
@@ -92,10 +98,12 @@ class ContentPageTests(TestCase):
     @mock.patch("home.models.create_whatsapp_template")
     def test_template_create_with_buttons_on_save(self, mock_create_whatsapp_template):
         page = create_page(is_whatsapp_template=True, has_quick_replies=True)
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_with(
             f"wa_title_{page.get_latest_revision().id}",
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             ["button 1", "button 2"],
             None,
             [],
@@ -107,10 +115,12 @@ class ContentPageTests(TestCase):
         self, mock_create_whatsapp_template
     ):
         page = create_page(is_whatsapp_template=True, add_example_values=True)
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_with(
             f"wa_title_{page.get_latest_revision().id}",
             "Test WhatsApp Message with two variables, {{1}} and {{2}}",
             "UTILITY",
+            en,
             [],
             None,
             [],
@@ -124,10 +134,12 @@ class ContentPageTests(TestCase):
         template name
         """
         page = create_page(is_whatsapp_template=True, has_quick_replies=True)
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_once_with(
             f"wa_title_{page.get_latest_revision().pk}",
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             ["button 1", "button 2"],
             None,
             [],
@@ -143,6 +155,7 @@ class ContentPageTests(TestCase):
             expected_title,
             "Test WhatsApp Message 2",
             "UTILITY",
+            en,
             ["button 1", "button 2"],
             None,
             [],
@@ -161,10 +174,12 @@ class ContentPageTests(TestCase):
         page.get_latest_revision().publish()
         page.refresh_from_db()
         expected_template_name = f"wa_title_{page.get_latest_revision().pk}"
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_once_with(
             expected_template_name,
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             ["button 1", "button 2"],
             None,
             [],
@@ -196,10 +211,12 @@ class ContentPageTests(TestCase):
         page.refresh_from_db()
         expected_template_name = f"wa_title_{page.get_latest_revision().pk}"
         self.assertEqual(page.whatsapp_template_name, expected_template_name)
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_once_with(
             expected_template_name,
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             ["button 1", "button 2"],
             None,
             [],
@@ -228,10 +245,12 @@ class ContentPageTests(TestCase):
         page.save_revision()
 
         expected_template_name = f"wa_title_{page.get_latest_revision().pk}"
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_once_with(
             expected_template_name,
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             [],
             None,
             [],
@@ -314,10 +333,12 @@ class ContentPageTests(TestCase):
         page = create_page(is_whatsapp_template=True)
         page.get_latest_revision().publish()
         expected_template_name = f"wa_title_{page.get_latest_revision().pk}"
+        en = Locale.objects.get(language_code="en")
         mock_create_whatsapp_template.assert_called_once_with(
             expected_template_name,
             "Test WhatsApp Message 1",
             "UTILITY",
+            en,
             [],
             None,
             [],
@@ -347,6 +368,93 @@ class ContentPageTests(TestCase):
             output.getvalue().strip(),
             "No changes detected",
             "There are missing migrations:\n %s" % output.getvalue(),
+        )
+
+    def test_get_all_links(self):
+        """
+        ContentPage.get_all_links() should return two lists with all ContentPage and
+        OrderedContentSet links.
+        """
+        home_page = HomePage.objects.first()
+        main_menu = PageBuilder.build_cpi(home_page, "main-menu", "Main Menu")
+        test_page = PageBuilder.build_cp(
+            parent=main_menu, slug="page1", title="Page1", bodies=[]
+        )
+        page_with_links = PageBuilder.build_cp(
+            parent=main_menu,
+            slug="page2",
+            title="Page2",
+            bodies=[
+                WABody(
+                    "Page2",
+                    [
+                        WABlk(
+                            "Page2 WA Body",
+                            buttons=[PageBtn("Import Export", page=test_page)],
+                        )
+                    ],
+                )
+            ],
+        )
+        page_with_links = PageBuilder.link_related(page_with_links, [test_page])
+
+        ocs = OrderedContentSet(name="Test set")
+        ocs.pages.append(("pages", {"contentpage": test_page}))
+        ocs.save()
+        ocs.save_revision().publish()
+
+        page_links, ocs_links = test_page.get_all_links()
+
+        self.assertListEqual(
+            [
+                (
+                    f"/admin/pages/{page_with_links.id}/edit/#tab-whatsapp",
+                    "Page2 - WhatsApp: Go to button",
+                ),
+                (
+                    f"/admin/pages/{page_with_links.id}/edit/#tab-promotional",
+                    "Page2 - Related Page",
+                ),
+            ],
+            page_links,
+        )
+        self.assertListEqual(
+            [(f"/admin/snippets/home/orderedcontentset/edit/{ocs.id}/", "Test set")],
+            ocs_links,
+        )
+
+    def test_get_all_links_no_links(self):
+        """
+        ContentPage.get_all_links() should return two empty lists if there are no links
+        """
+        home_page = HomePage.objects.first()
+        main_menu = PageBuilder.build_cpi(home_page, "main-menu", "Main Menu")
+        test_page = PageBuilder.build_cp(
+            parent=main_menu, slug="page1", title="Page1", bodies=[]
+        )
+
+        page_links, ocs_links = test_page.get_all_links()
+
+        self.assertListEqual([], page_links)
+        self.assertListEqual([], ocs_links)
+
+    @override_settings(WHATSAPP_CREATE_TEMPLATES=True)
+    @mock.patch("home.models.create_whatsapp_template")
+    @pytest.mark.xfail(
+        reason="This fails because we can't get locale to create the page, "
+        "these tests will be changed once whatsapp templates are separated."
+    )
+    def test_template_create_with_pt_language(self, mock_create_whatsapp_template):
+        page = create_page(is_whatsapp_template=True)
+        pt, _created = Locale.objects.get_or_create(language_code="pt")
+        mock_create_whatsapp_template.assert_called_with(
+            f"wa_title_{page.get_latest_revision().id}",
+            "Test WhatsApp Message 1",
+            "UTILITY",
+            pt,
+            [],
+            None,
+            [],
         )
 
 
@@ -529,8 +637,8 @@ class SMSBlockTests(TestCase):
 
     def test_clean_text_char_limit(self):
         """Text messages should be limited to 160 characters"""
-        SMSBlock().clean(self.create_message_value(message="a" * 160))
+        SMSBlock().clean(self.create_message_value(message="a" * 459))
 
         with self.assertRaises(StructBlockValidationError) as e:
-            SMSBlock().clean(self.create_message_value(message="a" * 161))
+            SMSBlock().clean(self.create_message_value(message="a" * 460))
         self.assertEqual(list(e.exception.block_errors.keys()), ["message"])