+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Increased SMS limit to 459 characters
+- Validate character limit fir footer and list items when importing
+- Add Document upload extension validation
 
 -->
 ## v1.1.0